Move all references to new errors
