+++ src/python/zigopt/conditionals/from_json.py
@@ -1,9 +1,10 @@
 # Copyright © 2022 Intel Corporation
 #
 # SPDX-License-Identifier: Apache License 2.0
+from zigopt.api.errors import InvalidValueError
 from zigopt.handlers.validate.experiment import validate_conditional_value
 from zigopt.handlers.validate.validate_dict import ValidationType, get_opt_with_validation, get_with_validation
-from zigopt.net.errors import BadParamError, InvalidValueError
+from zigopt.net.errors import BadParamError
 from zigopt.parameters.from_json import set_parameter_name_from_json
