+++ src/python/zigopt/handlers/experiments/create.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: Apache License 2.0
 from zigopt.common import *
 from zigopt.api.auth import api_token_authentication, client_token_authentication
+from zigopt.api.errors import InvalidValueError, MissingJsonKeyError
 from zigopt.common.sigopt_datetime import current_datetime
 from zigopt.common.struct import ImmutableStruct
 from zigopt.conditionals.from_json import set_experiment_conditionals_list_from_json
@@ -36,7 +37,7 @@ from zigopt.handlers.validate.validate_dict import (
   key_present,
 )
 from zigopt.json.builder import ExperimentJsonBuilder
-from zigopt.net.errors import BadParamError, InvalidValueError, MissingJsonKeyError
+from zigopt.net.errors import BadParamError
 from zigopt.parameters.from_json import set_experiment_parameter_list_from_json
 from zigopt.protobuf.gen.experiment.experimentmeta_pb2 import (
   PARAMETER_DOUBLE,
