+++ src/python/zigopt/handlers/users/update.py
@@ -3,13 +3,14 @@
 # SPDX-License-Identifier: Apache License 2.0
 from zigopt.common import *
 from zigopt.api.auth import api_token_authentication
+from zigopt.api.errors import MissingParamError
 from zigopt.common.struct import ImmutableStruct
 from zigopt.handlers.users.base import UserHandler
 from zigopt.handlers.validate.user import validate_user_email, validate_user_name, validate_user_password
 from zigopt.handlers.validate.validate_dict import ValidationType, get_opt_with_validation
 from zigopt.iam_logging.service import IamEvent, IamResponseStatus
 from zigopt.json.builder import UserJsonBuilder
-from zigopt.net.errors import BadParamError, ForbiddenError, MissingParamError, NotFoundError
+from zigopt.net.errors import BadParamError, ForbiddenError, NotFoundError
 from zigopt.user.model import do_password_hash_work_factor_update, password_matches
