+++ src/python/zigopt/handlers/validate/assignments.py
@@ -3,10 +3,11 @@
 # SPDX-License-Identifier: Apache License 2.0
 from typing import Any
 
+from zigopt.api.errors import InvalidValueError
 from zigopt.experiment.constraints import human_readable_constraint  # type: ignore
 from zigopt.experiment.model import Experiment, ExperimentParameterProxy  # type: ignore
 from zigopt.handlers.validate.validate_dict import ValidationType, validate_type
-from zigopt.net.errors import BadParamError, InvalidValueError, ServerError  # type: ignore
+from zigopt.net.errors import BadParamError, ServerError  # type: ignore
 from zigopt.protobuf.gen.experiment.experimentmeta_pb2 import ExperimentConstraint, ExperimentParameter  # type: ignore
 
 from libsigopt.aux.constant import ConstraintType  # type: ignore
