+++ test/python/sigopttest/handlers/validate/experiment.py
@@ -6,8 +6,8 @@ import copy
 
 import pytest
 
+from zigopt.api.errors import InvalidValueError
 from zigopt.handlers.validate.experiment import validate_experiment_json_dict_for_create
-from zigopt.net.errors import InvalidValueError
 
 
 OK_NAME = "name"
